Separate implementation from main process invocation. Put logic into cli.js and invoke from index.js
